grt/rcx: fix performance-type-promotion-in-math-fn

+++ src/rcx/include/rcx/extProgressTracker.hpp
@@ -15,8 +15,7 @@ class ExtProgressTracker
  public:
   // Constructor with optional custom progress interval (default 5%)
   explicit ExtProgressTracker(int totalWires, float progressInterval = 5.0f)
-      : totalWireCount_(totalWires),
-        progressInterval_(progressInterval)
+      : totalWireCount_(totalWires), progressInterval_(progressInterval)
   {
   }
 
@@ -72,8 +71,7 @@ class ExtProgressTracker
 
   float calculateCurrentPercent() const
   {
-    return ceil(100.0f * static_cast<float>(extractedWireCount_)
-                / totalWireCount_);
+    return std::ceil(100.0f * extractedWireCount_ / totalWireCount_);
   }
 
   bool shouldPrintProgress(float currentPercent) const
